:memo: Update Revire RatingScore Type
Update Revire RatingScore Type

+++ server/src/main/java/com/ssafy/live/account/realtor/controller/dto/RealtorResponse.java
@@ -176,7 +176,7 @@ public class RealtorResponse {
 
             private Long reviewNo;
             private String reviewInfo;
-            private int ratingScore;
+            private float ratingScore;
             private LocalDateTime createDate;
 
             public static Reviews toEntity(Review review) {
